Prevent TinyMCE from converting links with linktype attribute to anchors.
Wagtail uses linktype and id attributes for their links and render the
proper links in views. However, TinyMCE expects an actual href to be
provided, otherwise it'll force the link to become an anchor. To avoid
this, format_value looks up the actual object and set href and any other
attribute that would be useful for its chooser. This way we don't have
to touch TinyMCE code.